Simple scene mechanics completed
GameScene is not over.
There are 4 scenes right now.
WelcomeScene: Basically, a main menu, and if a user comes with a URL then a redirector.
OpponentWaitingScene: Temporarily waiting screen for the opponent.
GameCreatingScene: This scene takes screen size and gives the most optimal (suitable for both players' screens) game size.
GameScene: Game itself.